requirements.txt
pandas>=2.0.0
impyla>=0.18.0
thrift>=0.13.0
thrift-sasl>=0.4.3
pure-sasl>=0.6.2
requests>=2.28.0
openpyxl>=3.0.0
numpy>=1.23.0
streamlit>=1.20.0
fpdf2>=2.7.0
